connect buttons to modal
